requirements.txt
Django==5.1.2
coreapi==2.3.3
psycopg2==2.9.9
python-dotenv==1.0.1
djangorestframework==3.15.2